restructure project layer
