+++ points/app.py
@@ -3,15 +3,16 @@ import ast
 import sys
 import os
 
+
 curPath = os.path.abspath(os.path.dirname(__file__))
 rootPath = os.path.split(curPath)[0]
 sys.path.append(rootPath)
 
-from points.src.configuration.log_config import setup_logging
-from points.src.common.util import Util
-from points.src.exception.validation_exception import ValidationException
-from points.src.service.position_calculator import PositionCalculator
-from points.src.service.validation.plateau_validator import PlateauValidator
+from points.src.main.configuration.log_config import setup_logging
+from points.src.main.service.validation.plateau_validator import PlateauValidator
+from points.src.main.common.util import Util
+from points.src.main.exception.validation_exception import ValidationException
+from points.src.main.service.position_calculator import PositionCalculator
 
 __author__ = 'Jayden'
