+++ points/src/main/common/util.py
