+++ points/src/main/exception/validation_exception.py
