+++ points/src/main/service/position_calculator.py
@@ -1,10 +1,10 @@
 import logging
 
-from points.src.common.util import Util
-from points.src.model.plateau import Plateau
-from points.src.model.position import Position
-from points.src.model.rover import Rover
-from points.src.service.validation.position_validator import PositionValidator
+from points.src.main.common.util import Util
+from points.src.main.model.plateau import Plateau
+from points.src.main.model.position import Position
+from points.src.main.model.rover import Rover
+from points.src.main.service.validation.position_validator import PositionValidator
 
 
 class PositionCalculator:
