+++ points/src/main/service/validation/base_validator.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod
 
-from points.src.model.validation_result import ValidationResult
+from points.src.main.model.validation_result import ValidationResult
 
 
 class BaseValidator:
