+++ points/src/tests/test_service/test_plateau_validator.py
@@ -1,8 +1,8 @@
 import unittest
 
 from nose.tools import assert_equal
-from points.src.common.util import Util
-from points.src.service.validation.plateau_validator import PlateauValidator
+from points.src.main.common.util import Util
+from points.src.main.service.validation.plateau_validator import PlateauValidator
 
 
 class TestPlateauValidator(unittest.TestCase):
